stack configuration: add source field
Because we need the source image for a stack so that we know how to grab
its configuration files, this change adds a source to the stack
configuration object. In a "real" situation, this field would come from
somewhere else, such as the stack install or the stack object itself.

Signed-off-by: suskin <813571+suskin@users.noreply.github.com>