engine controller: wire up real job creation
Now that we have a job definition in the controller, this changeset adds
the functionality for kicking off that job. It will create one job per
configured "hook". Currently, there is no notion of listening for a crd
type declared in the stack configuration, but we'll want to do that in
the future. I've been exercising the code path so far by configuring my
stack to listen for helmchartinstall objects.

Signed-off-by: suskin <813571+suskin@users.noreply.github.com>