controller: add some thoughts about remaining effort
The controller was originally written to target the execution of a
template hook, so it's missing a nice separation between the setup and
render phase. I've added some thoughts about this, and some reminders
about which parts are remaining to demonstrate a full set of
functionality.

Also, since this is a proof of concept, the code is a bit rough, so I've made
notes to remind myself of the improvements I was thinking about when
imagining the polished version of the code.

Signed-off-by: suskin <813571+suskin@users.noreply.github.com>