controller: fix reading and writing of config map
Previously, the claim's spec contents were being truncated because they
didn't exist in the typed representation. We won't have access to a
typed representation anyway in a shared controller, so now the
controller uses an unstructured client to fetch the claim.

Also, the conversion of the byte array to a string was not working as
expected, so that's fixed too.

Signed-off-by: suskin <813571+suskin@users.noreply.github.com>