controller: remove processConfiguration method
This refactor removes a level of hierarchy from the method calls. The
`processConfiguration` method was acting as a top-level method and its
behavior didn't match its name, so it was making the code less readable.
This puts its flow logic into the render method, and puts its helper
logic into a helper method to pull out the behavior configuration.

Signed-off-by: suskin <813571+suskin@users.noreply.github.com>