setup controller: add first pass at dynamic registration
When we see a stack configuration, we'll want to set up registration for
the types mentioned in the configuration. This adds a proof of concept
of that behavior, by dynamically registering the known render phase type.
It's currently hard-coded, but it would normally be reading the type
from the stack configuration.

Signed-off-by: suskin <813571+suskin@users.noreply.github.com>