stack configuration: use structure more like template stacks design doc
We were using a stack configuration which had an outdated structure.
This updates the stack configuration to match the template stacks
experience design document; a bunch of other code has been refactored as
a result. A side effect of making this change is that the config map
creation has been updated to be idempotent, i.e. not fail when the
config map already exists.

Signed-off-by: suskin <813571+suskin@users.noreply.github.com>