render phase: add resource engine interface
As part of being able to use multiple different engines, we want to be
able to have a common contract across all resource engines (such as
helm2, helm3, kustomize). This change adds an initial interface for
that, and ports some of the helm2 engine functionality to be behind the
interface.

Signed-off-by: suskin <813571+suskin@users.noreply.github.com>