render controller: take config name on creation
Previously, the render controller was fetching the stack configuration
based on the name of the claim that it was reconciling. Now, it takes in
the name of the stack configuration when it's created, because an
individual stack controller always fetches the same stack configuration,
and grabbing the name from the claim was a bit of a temporary measure to
get things up and running.

Signed-off-by: suskin <813571+suskin@users.noreply.github.com>