helm2 engine: assume full relative resource path
Previously, the path to the resources was assumed to be within the
`resources` folder in the stack's contents. However, it seems more
natural to put the resources in the `resources` root when there's only
one set of resource files, so this changes the helm2 engine to assume
the `resources` directory is specified as part of the path.

Signed-off-by: suskin <813571+suskin@users.noreply.github.com>